+++ CHANGELOG.md
@@ -12,6 +12,7 @@
 - Added resource discovery support for `oce` resources
 - Support resource discovery for `oda` resources
 - Support resource discovery for `datascience` resources
+- Support resource discovery for `oci_objectstorage_object`, `oci_objectstorage_object_lifecycle_policy`, `oci_objectstorage_preauthrequest` resources
 
 ### Fixed
 - Fixed plan failure in case of missing required attributes in resource discovery. Placeholder values will be added for missing required attributes and the attributes will be added to `lifecycle ignore_changes`